Css update
Added the border for the table headers and the table data cells. Added the background color for the table data cells. Added the background color for the table data with the data date attribute. Added the Caption styles. Added padding, and font styles. Added the background color and the text color for the table row in the table heading.  Added the column gap added the column width, column rule, widows, orphans. Added the column span for the primary headings in the the article.